Use mixin to solve code duplication An issue subsists : not using a filter removes auto-reload of i18n
