+++ src/common/styles/cd-event-tile.less
@@ -1,0 +1,106 @@
+@import "../variables.less";
+
+.cd-event-tile() {
+  &__details {
+    display: flex;
+  }
+
+  &__header {
+    flex: 4;
+  }
+
+  &__name {
+    font-size: 24px;
+    margin: 0;
+  }
+
+  &__sessions {
+    font-size: 16px;
+    color: #7b8082;
+    margin: 8px 0;
+  }
+
+  &__date, &__times {
+    flex: 1;
+    &-timestamp {
+      list-style-type: none;
+      padding: 0;
+      font-size: 18px;
+    }
+  }
+
+  &__date {
+    &-timestamp {
+      font-size: 18px;
+    }
+
+    &-series {
+      font-size: 14px;
+      color: #7b8082;
+    }
+  }
+
+  &__times-timestamp {
+    font-size: 16px;
+  }
+
+  &__recurring-info {
+    background-color: #f4f5f6;
+    color: #7b8082;
+    margin: 32px -16px -16px -16px;
+    padding: 16px;
+    &-icon {
+      font-size: 16px;
+      margin-right: 4px;
+    }
+    &-header {
+      font-weight: bold;
+    }
+    &-text {
+      font-size: 14px;
+    }
+  }
+
+  &__view {
+    margin-top: 32px;
+    &:disabled {
+       background-color: #bdc3c6;
+       color: white;
+       font-weight: bold;
+       width: 220px;
+       text-transform: uppercase;
+       border-color: transparent;
+    }
+  }
+
+  &__datetime {
+    margin-right: 27px;
+    flex: 1;
+  }
+  }
+
+  @media (max-width: @screen-xs-max) {
+  .cd-event-list-item {
+    &__details {
+      flex-direction: column;
+    }
+    &__name {
+      font-size: 18px;
+      font-weight: bold;
+    }
+    &__sessions {
+      font-size: 14px;
+    }
+    &__view-wrapper {
+      text-align: center;
+    }
+    &__recurring-info {
+      width: auto;
+    }
+    &__datetime {
+      margin-right: 0;
+      margin-top: 16px;
+      text-align: center;
+    }
+  }
+}
