+++ test/unit/specs/dojos/cd-dojo-details.spec.js
@@ -93,14 +93,15 @@ describe('Dojo details component', () => {
         id: '74afa4b8-8449-46e4-a553-8febda8614ad',
       };
       vm.dojoDetails = {
-        id: '3ed47c6d-a689-46a0-883b-1f3fd46e9c77'
+        id: '3ed47c6d-a689-46a0-883b-1f3fd46e9c77',
       };
 
       // ACT
       vm.volunteer(userType);
 
       // ASSERT
-      expect(DojoServiceMock.requestUserInvite).to.have.been.calledWith(vm.user, vm.dojoDetails.id, userType);
+      expect(DojoServiceMock.requestUserInvite).to
+        .have.been.calledWith(vm.user, vm.dojoDetails.id, userType);
     });
   });
